[stable20.1] Don't remove a chat when a self-joined user leaves

+++ lib/Controller/RoomController.php
@@ -1309,7 +1309,12 @@ class RoomController extends AEnvironmentAwareController {
 
 		if ($room->getType() !== Room::CHANGELOG_CONVERSATION &&
 			$room->getObjectType() !== 'file' &&
-			$this->participantService->getNumberOfUsers($room) === 1) {
+			$this->participantService->getNumberOfUsers($room) === 1 &&
+			\in_array($participant->getAttendee()->getParticipantType(), [
+				Participant::USER,
+				Participant::MODERATOR,
+				Participant::OWNER,
+			], true)) {
 			$room->deleteRoom();
 			return new DataResponse();
 		}
